Feature: Quickbuy Executor
It is now possible to actually buy something.
The quickbuy executor takes the information from the quickbuy parser.
It does these steps:
1. Validate username and get user id from it
2. Calculate how much money the user has
3. Check the products and get their ids
4. Add a transaction for each product for each amount of the product
5. ???
6. Profit

We still need unit tests for this but I don't know yet how to do this
without abstracting everything to high hell.

The parser and executer still needs to provide multiple errors and
exactly where the error occurred in the commandline string.

Implements #5

Co-authored-by: Casper-NS <caspernyvang@gmail.com>