UI: Use templating system
Now we use a templating system for html that looks alot like the one in
the python stregsystem.

The basic parts of the html is done (header, footer).

Made the ui a bit nicer.

Made the quickbuy buy button function (no response on the UI look in the
console output).