Implement opt-in idempotency for non-idempotent http methods
The client can provide a `X-Idempotency-Key` header with a random string
as a value.
All requests with this idempotency key will only be executed once
returning cached response if key is seen again.
This allows for bad-network scenarios where a response from the server
is lost and client sends a new request with same idempotency key.
The server already handled the request and as such just the cached
response is sent back.

This is useful for purchase requests where multiple purchases are not
wanted.

Currently the following http methods are handled: `POST` and `PATCH`.

If no idempotency key is given the idempotency system is bypassed.

The cache is a Least-Recently-Used cache with 256 entries.
You can probably see the issue if a to-be-used idempotency key is
evicted from the cache as the requests would be handled again by the
server.
This is a trade off.
Tune the size value as needed.